Make breacrumbs show Automation Server -> Workspace -> Automation (#113)
